requirements.txt
Django              5.0.6
djangorestframework 3.15.1